Add files via upload
Weather is the state of the atmosphere at a given place, date and time in regards to heat, cloudiness, dryness, sunshine, wind, and rain.
 Of all the geophysical phenomena weather is the most significant one that influences us. Weather can vary greatly and largely depends on climate, seasons and various 
     other factors. 
The chief goal of this work is to get the weather forecast of any city throughout the world through an application. 
This paper aims at creating a weather application using JavaScript for fetching the API, Using CSS for Styling and HTML used for Structure for Whether app.